check if product exists in database before deleting

+++ src/main/java/com/latebox/server/product/ProductController.java
@@ -49,7 +49,11 @@ public class ProductController {
 
     @DeleteMapping("/products/{id}")
     void deleteProduct(@PathVariable Long id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)){
+            repository.deleteById(id);
+        }else{
+            throw new ProductNotFoundException(id);
+        }
     }
 }
